[somfytahoma] Made constants and handler packages internal
See eclipse/smarthome#4330
