+++ addons/binding/org.openhab.binding.somfytahoma/META-INF/MANIFEST.MF
@@ -28,10 +28,6 @@ Import-Package: com.google.gson,
  org.eclipse.smarthome.core.thing.binding.builder,
  org.eclipse.smarthome.core.thing.type,
  org.eclipse.smarthome.core.types,
- org.openhab.binding.somfytahoma,
- org.openhab.binding.somfytahoma.handler,
  org.osgi.framework,
  org.slf4j
 Service-Component: OSGI-INF/*.xml
-Export-Package: org.openhab.binding.somfytahoma,
- org.openhab.binding.somfytahoma.handler
