+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaAwningHandler.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
@@ -19,56 +19,55 @@ import org.eclipse.smarthome.core.types.RefreshType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants.*;
+
 import java.util.HashMap;
 
-import static org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants.*;
-
 /**
- * The {@link SomfyTahomaVenetianBlindHandler} is responsible for handling commands,
- * which are sent to one of the channels of the venetian blind thing.
+ * The {@link SomfyTahomaAwningHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the awning thing.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaVenetianBlindHandler extends SomfyTahomaBaseThingHandler {
+public class SomfyTahomaAwningHandler extends SomfyTahomaBaseThingHandler {
 
-    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaVenetianBlindHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaAwningHandler.class);
 
-    public SomfyTahomaVenetianBlindHandler(Thing thing) {
+    public SomfyTahomaAwningHandler(Thing thing) {
         super(thing);
         stateNames = new HashMap<String, String>() {{
-            put(CONTROL, "core:ClosureState");
-            put(ORIENTATION, "core:SlateOrientationState");
+            put(CONTROL, "core:DeploymentState");
         }};
     }
 
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
         logger.debug("Received command {} for channel {}", command, channelUID);
-        if (!CONTROL.equals(channelUID.getId()) && !ORIENTATION.equals(channelUID.getId())) {
+        if (!CONTROL.equals(channelUID.getId())) {
             return;
         }
 
         if (RefreshType.REFRESH.equals(command)) {
             updateChannelState(channelUID);
         } else {
-            String cmd = getTahomaCommand(command.toString(), channelUID.getId());
+            String cmd = getTahomaCommand(command.toString());
             if (COMMAND_MY.equals(cmd)) {
                 String executionId = getCurrentExecutions();
                 if (executionId != null) {
-                    //Check if the venetian blind is moving and MY is sent => STOP it
+                    //Check if the awning is moving and MY is sent => STOP it
                     cancelExecution(executionId);
                 } else {
                     sendCommand(COMMAND_MY, "[]");
                 }
             } else {
-                String param = (COMMAND_SET_CLOSURE.equals(cmd) || COMMAND_SET_ORIENTATION.equals(cmd)) ? "[" + command.toString() + "]" : "[]";
+                String param = COMMAND_SET_DEPLOYMENT.equals(cmd) ? "[" + command.toString() + "]" : "[]";
                 sendCommand(cmd, param);
             }
         }
 
     }
 
-    private String getTahomaCommand(String command, String channelId) {
+    private String getTahomaCommand(String command) {
         switch (command) {
             case "OFF":
             case "DOWN":
@@ -79,11 +78,7 @@ public class SomfyTahomaVenetianBlindHandler extends SomfyTahomaBaseThingHandler
             case "STOP":
                 return COMMAND_MY;
             default:
-                if (CONTROL.equals(channelId)) {
-                    return COMMAND_SET_CLOSURE;
-                } else {
-                    return COMMAND_SET_ORIENTATION;
-                }
+                return COMMAND_SET_DEPLOYMENT;
         }
     }
 }
