+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaContactSensorHandler.java
@@ -10,44 +10,43 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
-import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
-import org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants.*;
+
 import java.util.HashMap;
 
-import static org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants.SWITCH;
-
 /**
- * The {@link SomfyTahomaOnOffHandler} is responsible for handling commands,
- * which are sent to one of the channels of the on/off thing.
+ * The {@link SomfyTahomaContactSensorHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the contact sensors.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaOnOffHandler extends SomfyTahomaBaseThingHandler {
+public class SomfyTahomaContactSensorHandler extends SomfyTahomaBaseThingHandler  {
 
-    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaOnOffHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaContactSensorHandler.class);
 
-    public SomfyTahomaOnOffHandler(Thing thing) {
+    public SomfyTahomaContactSensorHandler(Thing thing) {
         super(thing);
         stateNames = new HashMap<String, String>() {{
-            put(SWITCH, "core:OnOffState");
+            put(CONTACT, "core:ContactState");
         }};
     }
 
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
         logger.debug("Received command {} for channel {}", command, channelUID);
-        if (SomfyTahomaBindingConstants.SWITCH.equals(channelUID.getId()) && command instanceof OnOffType) {
-            sendCommand(command.toString().toLowerCase(), "[]");
+        if (!CONTACT.equals(channelUID.getId())) {
+            return;
         }
+
         if (RefreshType.REFRESH.equals(command)) {
             updateChannelState(channelUID);
         }
