+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaHeatingSystemHandler.java
@@ -10,7 +10,7 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
@@ -19,25 +19,28 @@ import org.eclipse.smarthome.core.types.RefreshType;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants.*;
+
 import java.util.HashMap;
 
 /**
- * The {@link SomfyTahomaPodHandler} is responsible for handling commands,
- * which are sent to one of the channels of the pod thing.
+ * The {@link SomfyTahomaHeatingSystemHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the heating system thing.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaPodHandler extends SomfyTahomaBaseThingHandler {
+public class SomfyTahomaHeatingSystemHandler extends SomfyTahomaBaseThingHandler {
 
-    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaPodHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaHeatingSystemHandler.class);
 
-    public SomfyTahomaPodHandler(Thing thing) {
+    public SomfyTahomaHeatingSystemHandler(Thing thing) {
         super(thing);
         stateNames = new HashMap<String, String>() {
             {
-                put("cyclic_button_state", "core:CyclicButtonState");
-                put("battery_status_state", "internal:BatteryStatusState");
-                put("lighting_led_pod_mod_state", "internal:LightingLedPodModeState");
+                put(TARGET_TEMPERATURE, "core:TargetTemperatureState");
+                put(CURRENT_TEMPERATURE, "zwave:SetPointHeatingValueState");
+                put(BATTERY_LEVEL, "core:BatteryLevelState");
+                put(CURRENT_STATE, "zwave:SetPointTypeState");
             }
         };
     }
@@ -47,6 +50,11 @@ public class SomfyTahomaPodHandler extends SomfyTahomaBaseThingHandler {
         logger.debug("Received command {} for channel {}", command, channelUID);
         if (RefreshType.REFRESH.equals(command)) {
             updateChannelState(channelUID);
+        } else {
+            if (TARGET_TEMPERATURE.equals(channelUID.getId())) {
+                String param = "[" + command.toString() + "]";
+                sendCommand("setTargetTemperature", param);
+            }
         }
     }
 }
