+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaOnOffHandler.java
@@ -10,43 +10,44 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
+import org.eclipse.smarthome.core.library.types.OnOffType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
 import org.eclipse.smarthome.core.types.RefreshType;
+import org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import static org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants.SWITCH;
+
 import java.util.HashMap;
 
-import static org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants.LUMINANCE;
-
 /**
- * The {@link SomfyTahomaLightSensorHandler} is responsible for handling commands,
- * which are sent to one of the channels of the light sensor thing.
+ * The {@link SomfyTahomaOnOffHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the on/off thing.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaLightSensorHandler extends SomfyTahomaBaseThingHandler {
+public class SomfyTahomaOnOffHandler extends SomfyTahomaBaseThingHandler {
 
-    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaLightSensorHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaOnOffHandler.class);
 
-    public SomfyTahomaLightSensorHandler(Thing thing) {
+    public SomfyTahomaOnOffHandler(Thing thing) {
         super(thing);
         stateNames = new HashMap<String, String>() {{
-            put(LUMINANCE, "core:LuminanceState");
+            put(SWITCH, "core:OnOffState");
         }};
     }
 
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
         logger.debug("Received command {} for channel {}", command, channelUID);
-        if (!LUMINANCE.equals(channelUID.getId())) {
-            return;
+        if (SomfyTahomaBindingConstants.SWITCH.equals(channelUID.getId()) && command instanceof OnOffType) {
+            sendCommand(command.toString().toLowerCase(), "[]");
         }
-
         if (RefreshType.REFRESH.equals(command)) {
             updateChannelState(channelUID);
         }
