+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaPodHandler.java
@@ -10,9 +10,8 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
-import org.eclipse.smarthome.core.library.types.StringType;
 import org.eclipse.smarthome.core.thing.ChannelUID;
 import org.eclipse.smarthome.core.thing.Thing;
 import org.eclipse.smarthome.core.types.Command;
@@ -22,35 +21,32 @@ import org.slf4j.LoggerFactory;
 
 import java.util.HashMap;
 
-import static org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants.ALARM_COMMAND;
-
 /**
- * The {@link SomfyTahomaExternalAlarmHandler} is responsible for handling commands,
- * which are sent to one of the channels of the alarm thing.
+ * The {@link SomfyTahomaPodHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the pod thing.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaExternalAlarmHandler extends SomfyTahomaBaseThingHandler {
+public class SomfyTahomaPodHandler extends SomfyTahomaBaseThingHandler {
 
-    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaExternalAlarmHandler.class);
+    private final Logger logger = LoggerFactory.getLogger(SomfyTahomaPodHandler.class);
 
-    public SomfyTahomaExternalAlarmHandler(Thing thing) {
+    public SomfyTahomaPodHandler(Thing thing) {
         super(thing);
-        stateNames = new HashMap<String, String>() {{
-            put("active_zones_state", "core:ActiveZonesState");
-        }};
+        stateNames = new HashMap<String, String>() {
+            {
+                put("cyclic_button_state", "core:CyclicButtonState");
+                put("battery_status_state", "internal:BatteryStatusState");
+                put("lighting_led_pod_mod_state", "internal:LightingLedPodModeState");
+            }
+        };
     }
 
     @Override
     public void handleCommand(ChannelUID channelUID, Command command) {
         logger.debug("Received command {} for channel {}", command, channelUID);
-        if (ALARM_COMMAND.equals(channelUID.getId()) && command instanceof StringType) {
-            sendCommand(command.toString(), "[]");
-        }
         if (RefreshType.REFRESH.equals(command)) {
-            sendCommand("refreshState", "[]");
             updateChannelState(channelUID);
         }
     }
-
 }
