+++ addons/binding/org.openhab.binding.somfytahoma/src/main/java/org/openhab/binding/somfytahoma/internal/handler/SomfyTahomaSmokeSensorHandler.java
@@ -10,27 +10,26 @@
  *
  * SPDX-License-Identifier: EPL-2.0
  */
-package org.openhab.binding.somfytahoma.handler;
+package org.openhab.binding.somfytahoma.internal.handler;
 
 import org.eclipse.smarthome.core.thing.Thing;
 
+import static org.openhab.binding.somfytahoma.internal.SomfyTahomaBindingConstants.CONTACT;
+
 import java.util.HashMap;
 
-import static org.openhab.binding.somfytahoma.SomfyTahomaBindingConstants.CONTACT;
-
 /**
- * The {@link SomfyTahomaOccupancySensorHandler} is responsible for handling commands,
- * which are sent to one of the channels of the occupancy sensor thing.
+ * The {@link SomfyTahomaSmokeSensorHandler} is responsible for handling commands,
+ * which are sent to one of the channels of the smoke sensor thing.
  *
  * @author Ondrej Pecta - Initial contribution
  */
-public class SomfyTahomaOccupancySensorHandler extends SomfyTahomaContactSensorHandler {
+public class SomfyTahomaSmokeSensorHandler extends SomfyTahomaContactSensorHandler {
 
-    public SomfyTahomaOccupancySensorHandler(Thing thing) {
+    public SomfyTahomaSmokeSensorHandler(Thing thing) {
         super(thing);
         stateNames = new HashMap<String, String>() {{
-            put(CONTACT, "core:OccupancyState");
+            put(CONTACT, "core:SmokeState");
         }};
     }
-
 }
